https://www.file.io/uabH/download/qEHWfEi8jmT2
As mentioned earlier I'm new to coding really going back and forth in terms of learning to code and understanding the technical stuff. 
But here's a neat Binary to Text Converter which is really cool to play with if you're into that Geeky Stuff! Props to AI and Co-Pilot for helping me get started.
Nuff Said hope you enjoy it as much as I do!